% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_at.R
\name{delete_at}
\alias{delete_at}
\title{Delete at position(s)}
\usage{
delete_at(x, y)
}
\arguments{
\item{x}{vector/list}

\item{at}{integer/numeric}
}
\description{
Delete at position(s)
}
\examples{
x <- c(1, 2, 3, 4, 5, 0)
delete_at(x, 5)
delete_at(x, 4:5)
}
