% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_if.R
\name{delete_if}
\alias{delete_if}
\title{Delete if fun evaluates to TRUE}
\usage{
delete_if(x, fun)
}
\arguments{
\item{x}{list/vector}

\item{fun}{any function - has to yield a boolean}
}
\description{
Delete if fun evaluates to TRUE
}
\examples{
x <- c(1, 2, 3, 4, 5, 0)
delete_if(x, function(z) z > 2)
delete_if(x, function(z) z < 4)
}
