% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rubfuns-package.R
\docType{package}
\name{rubfuns-package}
\alias{rubfuns-package}
\alias{rubfuns}
\title{Interface to the ftdoi.org API for publisher url patterns.}
\description{
Interface to the ftdoi.org API for publisher url patterns.
}
\keyword{package}
