% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift.R
\name{shift}
\alias{shift}
\alias{unshift}
\title{Move element to front or back of a list/vector}
\usage{
shift(x, y = NULL)

unshift(x, y)
}
\arguments{
\item{x}{(list/vector) a list or vector}

\item{y}{(integer) a single integer. treated as a sequence of
numbers up to that number, starting at 1 (since R uses 1 based
indexing)}
}
\description{
Move element to front or back of a list/vector
}
\section{\code{shift}}{

Removes the first element of \code{x} and returns \code{x}. Returns \code{NULL}
if the list/vector is empty.

If a number \code{y} is given, returns an array of the first n elements
(or less) just like array.slice!(0, n) does. With ary containing
only the remainder elements, not including what was shifted to
new_ary
}

\section{\code{unshift}}{

Prepends objects to the front of self, moving other elements
upwards
}

\examples{
# shift
x <- c(1, 2, 3)
shift(x)
shift(x, 1)

# unshift
x <- c(1, 2, 3)
unshift(x, 4)
}
