% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take_while.R
\name{take_while}
\alias{take_while}
\alias{drop_while}
\title{Take while fun evaluates to TRUE}
\usage{
take_while(x, fun)

drop_while(x, fun)
}
\arguments{
\item{x}{(list/vector) a list or vector}

\item{fun}{(function) any function - has to yield a boolean}
}
\description{
Take while fun evaluates to TRUE
}
\examples{
# take while
x <- c(1, 2, 3, 4, 5, 0)
take_while(x, function(z) z < 3)
x <- c(1, 2, 3, 4, 9, -1)
take_while(x, function(z) z < 3)

# drop while
x <- c(1, 2, 3, 4, 5, 0)
drop_while(x, function(z) z < 3)
x <- c(1, 2, 3, 4, 9, -1)
drop_while(x, function(z) z < 3)
}
