% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take_while.R
\name{take_while}
\alias{take_while}
\title{Take while fun evaluates to TRUE}
\usage{
take_while(x, fun)
}
\arguments{
\item{x}{xxxxx}

\item{fun}{any function - has to yield a boolean}
}
\description{
Take while fun evaluates to TRUE
}
\examples{
# a = [1, 2, 3, 4, 5, 0]
# a.take_while {|i| i < 3 }   #=> [1, 2]

x <- c(1, 2, 3, 4, 5, 0)
take_while(x, function(z) z < 3)

x <- c(1, 2, 3, 4, 9, -1)
take_while(x, function(z) z < 3)

x <- c(1, 2, 3, 4, 9)
take_while(x, function(z) z < 3) \%>\% sqrt
}
